+++ server/controller.go
@@ -70,6 +70,7 @@ type Controller struct {
 	AlertEngine                      *AlertEngine
 	HallucinationDetector            *HallucinationDetector
 	CentralManagement                *CentralManagementService
+	Health                           *HealthService
 	// Performance caches
 	PreferencesCache  *PreferencesCache
 	KeywordListsCache *KeywordListsCache
@@ -194,6 +195,7 @@ func NewController(config *Config) *Controller {
 	controller.DeviceTokens = NewDeviceTokens()
 	controller.EmailService = NewEmailService(controller)
 	controller.CentralManagement = NewCentralManagementService(controller)
+	controller.Health = NewHealthService(controller)
 	controller.Delayer = NewDelayer(controller)
 	controller.Downstreams = NewDownstreams(controller)
 	controller.Scheduler = NewScheduler(controller)